+++ src/breakpoint_twa.hpp
@@ -41,7 +41,7 @@ class bp_twa {
         powerset_on_cut_ = om->get("powerset-on-cut",1);
         jump_to_bottommost_ = om->get("jump-to-bottommost",1);
         skip_levels_ = om->get("skip-levels",1);
-        reuse_SCC_ = om->get("reuse-good-SCC",1);
+        reuse_SCC_ = om->get("reuse-deterministic",1);
         cut_always_ = om->get("cut-always",0);
         cut_on_SCC_entry_ = om->get("cut-on-SCC-entry",0);
         bscc_avoid_ = (om->get("bscc-avoid", 1) || reuse_SCC_) ?