+++ src/main.cpp
@@ -78,14 +78,14 @@ Cut-edges construction:
 Optimizations:
     --bscc-avoid[=0|1]          avoid deterministic bottom part of input in 1st
                                 component and jump directly to 2nd component
+    --jump-to-bottommost[=0|1]  remove useless trivial SCCs of 2nd component
     --powerset-for-weak[=0|1]   avoid breakpoint construction for
                                 inherently weak accepting SCCs and use
                                 powerset construction instead
     --powerset-on-cut[=0|1]     if s -a-> p needs a cut, create
                                 s -a-> (δ(s),δ_0(s),0) instead of
                                 s -a-> ({p},∅,0).
-    --jump-to-bottommost[=0|1]     remove useless prefixes of second component
-    --reuse-good-SCC[=0|1]      similar as --bscc-avoid, but uses the SCCs
+    --reuse-deterministic[=0|1] similar to --bscc-avoid, but uses the SCCs
                                 unmodified with (potentialy) TGBA acceptance
     --skip-levels[=0|1]         allow multiple breakpoints on 1 edge; a trick
                                 well known from degeneralization
@@ -175,7 +175,7 @@ int main(int argc, char* argv[])
         else if (match_opt(arg, "--powerset-for-weak")
                  || match_opt(arg, "--jump-to-bottommost")
                  || match_opt(arg, "--bscc-avoid")
-                 || match_opt(arg, "--reuse-good-SCC")
+                 || match_opt(arg, "--reuse-deterministic")
                  || match_opt(arg, "--skip-levels")
                  || match_opt(arg, "--scc-aware")
                  || match_opt(arg, "--powerset-on-cut")
@@ -191,7 +191,7 @@ int main(int argc, char* argv[])
           {
             om.set("bscc-avoid", false);
             om.set("powerset-for-weak", false);
-            om.set("reuse-good-SCC", false);
+            om.set("reuse-deterministic", false);
             om.set("jump-to-bottommost", false);
             om.set("bscc-avoid", false);
             om.set("skip-levels", false);